Shorten hash to 16 bit

+++ src/context/block_unit.rs
@@ -1826,8 +1826,10 @@ impl ContextWriter<'_> {
     w.bit(marker as u16);
     bits += 1;
     if marker {
+      //use log::info;
+      //info!("Used a hash");
       // PERF: We can encode this in a more efficient manner
-      for byte in (hash as u32).to_be_bytes() {
+      for byte in (hash as u16).to_be_bytes() {
         w.bit(((byte >> 7) & 0b1).into());
         w.bit(((byte >> 6) & 0b1).into());
         w.bit(((byte >> 5) & 0b1).into());
